requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
pillow==11.3.0
python-multipart==0.0.20
werkzeug==3.1.3
requests==2.32.3
# OpenCV and NumPy required by table_cropper
opencv-python==4.12.0.88
numpy==2.2.6
